gestione dei ruoli

+++ Caravan.Common/Security/ICaravanSecurityRepository.cs
@@ -240,7 +240,7 @@ namespace Finsa.Caravan.Common.Security
         /// <exception cref="ArgumentException">
         ///   <paramref name="appName"/> or <paramref name="roleId"/> are null or empty.
         /// </exception>
-        Task<SecRole> GetRoleByIdAsync(string appName, long roleId);
+        Task<SecRole> GetRoleByIdAsync(string appName, int roleId);
 
         /// <summary>
         ///   </summary>
